+++ project2/controllers/authcontroller.js
@@ -13,7 +13,7 @@ exports.signin = function(req, res) {
 }
 exports.dashboard = function(req, res) {
  
-    res.render('dashboard');
+    res.render('dashboard-owner');
  
 }
 exports.logout = function(req, res) {
