+++ project2/routes/auth.js
@@ -1,30 +1,30 @@
-var authController = require('../controllers/authcontroller.js');
+// var authController = require('../controllers/authcontroller.js');
  
-module.exports = function(app, passport) {
+// module.exports = function(app, passport) {
  
-    app.get('/signup', authController.signup);
-    app.get('/signin', authController.signin);
-    app.post('/signup', passport.authenticate('local-signup', {
-        successRedirect: '/dashboard-owner',
+//     app.get('/signup', authController.signup);
+//     app.get('/signin', authController.signin);
+//     app.post('/signup', passport.authenticate('local-signup', {
+//         successRedirect: '/dashboard-owner',
  
-        failureRedirect: '/signup'
-    }
-    ));
-    app.get('/dashboard',isLoggedIn, authController.dashboard);
-    app.get('/logout',authController.logout);
-    app.post('/signin', passport.authenticate('local-signin', {
-        successRedirect: '/dashboard-owner',
+//         failureRedirect: '/signup'
+//     }
+//     ));
+//     app.get('/dashboard',isLoggedIn, authController.dashboard);
+//     app.get('/logout',authController.logout);
+//     app.post('/signin', passport.authenticate('local-signin', {
+//         successRedirect: '/dashboard-owner',
  
-        failureRedirect: '/signin'
-    }
+//         failureRedirect: '/signin'
+//     }
  
-));
-}
-function isLoggedIn(req, res, next) {
-    if (req.isAuthenticated())
+// ));
+// }
+// function isLoggedIn(req, res, next) {
+//     if (req.isAuthenticated())
      
-        return next();
+//         return next();
          
-    res.redirect('/signin');
+//     res.redirect('/signin');
  
-}
+// }
