+++ project2/routes/htmlRoutes.js
@@ -1,15 +1,33 @@
 var db = require("../models");
+var authController = require('../controllers/authcontroller.js');
 
-module.exports = function(app) {
+module.exports = function(app, passport) {
   // Load index page
   app.get("/", function(req, res) {
     db.Players.findAll({}).then(function(Players) {
-      res.render("signup.hbs", {
+      res.render("signup", {
         msg: "Welcome!",
         examples: Players
       });
     });
   });
+  app.get('/signup', authController.signup);
+  app.get('/signin', authController.signin);
+  app.post('/signup', passport.authenticate('local-signup', {
+      successRedirect: '/dashboard-owner',
+
+      failureRedirect: '/signup'
+  }
+  ));
+  app.get('/dashboard-owner',isLoggedIn, authController.dashboard);
+  app.get('/logout',authController.logout);
+  app.post('/signin', passport.authenticate('local-signin', {
+      successRedirect: '/dashboard-owner',
+
+      failureRedirect: '/signin'
+  }
+
+));
   // Load Dashboard page
   app.get("/dashboard-owner", function(req, res) {
     db.owners.findAll({}).then(function(dbPlayers) {
@@ -78,3 +96,11 @@ module.exports = function(app) {
     res.render("404");
   });
 };
+function isLoggedIn(req, res, next) {
+  if (req.isAuthenticated())
+   
+      return next();
+       
+  res.redirect('/signin');
+
+}
