+++ project2/server.js
@@ -40,8 +40,8 @@ var models = require("./models");
 
 // Routes
 require("./routes/apiRoutes")(app);
-require("./routes/htmlRoutes")(app);
-var authRoute = require('./routes/auth')(app, passport);
+require("./routes/htmlRoutes")(app, passport);
+// var authRoute = require('./routes/auth')(app, passport);
 
 var syncOptions = { force: false };
